feat: creating useGenerateUsername hook
